initial commit save 3

+++ src/merutilm/rff/settings/ReferenceCompressionSettings.h
@@ -3,10 +3,9 @@
 //
 
 #pragma once
-#include <cstdint>
 
 struct ReferenceCompressionSettings final{
-    const uint8_t compressCriteria;
+    const uint32_t compressCriteria;
     const uint8_t compressionThresholdPower;
